+++ src/main/java/no/digipat/patornat/servlets/NextImagePairServlet.java
@@ -40,7 +40,7 @@ public class NextImagePairServlet extends HttpServlet {
     /**
      * Gets a pair of images for comparison. The response body will contain
      * a JSON object of the form <code>{"pair": [id1, id2]}</code>, where {@code id1}
-     * and {@code id2} are integers.
+     * and {@code id2} are longs.
      * 
      * @param request the HTTP request
      * @param response the HTTP response
@@ -76,7 +76,7 @@ public class NextImagePairServlet extends HttpServlet {
     
     private static JSONObject createRequestJson(List<Image> images, List<BestImageChoice> comparisons) {
         JSONObject json = new JSONObject();
-        List<Integer> imageIds = new ArrayList<>();
+        List<Long> imageIds = new ArrayList<>();
         for (Image image : images) {
             imageIds.add(image.getId());
         }
